Move footer to partials/
