+++ src/web-ui/src/partials/Footer/Footer.vue
